# table: Improve Table to only render visible range columns for performance up to stable on 120 FPS