--- a/crates/story/src/table_story.rs
+++ b/crates/story/src/table_story.rs
@@ -301,17 +301,13 @@ impl TableDelegate for StockTableDelegate {
         }
     }
 
-    fn col_width(&self, col_ix: usize, _: &AppContext) -> Option<Pixels> {
-        if let Some(_) = self.columns.get(col_ix) {
-            if col_ix < 10 {
-                Some(120.0.into())
-            } else if col_ix < 20 {
-                Some(80.0.into())
-            } else {
-                Some(130.0.into())
-            }
+    fn col_width(&self, col_ix: usize, _: &AppContext) -> Pixels {
+        if col_ix < 10 {
+            120.0.into()
+        } else if col_ix < 20 {
+            80.0.into()
         } else {
-            None
+            130.0.into()
         }
     }